requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1 